# Conflicts:
#	.classpath
#	src/org/openedit/store/Category.java
#	src/org/openedit/store/DiscountCalculator.java
#	src/org/openedit/store/MoneyCalculator.java
#	src/org/openedit/store/Product.java
#	src/org/openedit/store/ProductAdder.java
#	src/org/openedit/store/Store.java
#	src/org/openedit/store/adjustments/SecurityGroupAdjustment.java
#	src/org/openedit/store/customer/Customer.java
#	src/org/openedit/store/gateway/StripeOrderProcessor.java
#	src/org/openedit/store/modules/CartModule.java
#	src/org/openedit/store/orders/Order.java
#	src/org/openedit/store/search/ProductLuceneIndexAll.java
#	src/org/openedit/store/search/ProductLuceneIndexer.java
#	src/org/openedit/store/search/ProductLuceneSearcher.java
#	webapp/WEB-INF/base/store/src/plugin.xml

+++ src/org/openedit/store/CartItem.java
@@ -27,7 +27,7 @@ import org.openedit.store.orders.Shipment;
  */
 public class CartItem extends BaseData
 {	
-	protected int fieldQuantity = 1;
+	protected double fieldQuantity = 1;
 	protected Money fieldYourPrice;
 	protected Money fieldWholesalePrice;
 	protected InventoryItem fieldInventoryItem;
@@ -78,7 +78,7 @@ public class CartItem extends BaseData
 		{
 			return null;
 		}
-		int quantity = getQuantity();
+		double quantity = getQuantity();
 		Money price = Money.ZERO;
 		Money priceByQuantity = getInventoryItem().getWholesalePriceByQuantity(quantity);
 		if (priceByQuantity != null)
@@ -112,7 +112,7 @@ public class CartItem extends BaseData
 		{
 			return null;
 		}
-		int quantity = getQuantity();
+		double quantity = getQuantity();
 		Money price = Money.ZERO;
 		Money priceByQuantity = getInventoryItem().getYourPriceByQuantity(quantity);
 		if (priceByQuantity != null)
@@ -135,12 +135,12 @@ public class CartItem extends BaseData
 	{
 		fieldYourPrice = inMoney;
 	}
-	public int getQuantity()
+	public double getQuantity()
 	{
 		return fieldQuantity;
 	}
 
-	public void setQuantity(int quantity) throws StoreException
+	public void setQuantity(double quantity) throws StoreException
 	{
 		fieldQuantity = quantity;
